shopping cart

+++ src/main/java/com/koowakchai/store/service/TTotalOrderService.java
@@ -3,6 +3,6 @@ package com.koowakchai.store.service;
 
 public interface TTotalOrderService {
 
-    public long addTTotalOrderEntity(long cartEntityId, String remark) throws Exception;
+    public long addTTotalOrderEntity(int cartEntityId, String remark) throws Exception;
 
 }
